+++ src/display/mod.rs
@@ -30,6 +30,7 @@ pub mod flags_display;
 pub mod status_line;
 mod matched_string;
 mod screen;
+mod cell_size;
 
 #[cfg(not(any(target_family="windows",target_os="android")))]
 mod permissions;
@@ -43,6 +44,7 @@ pub use {
     git_status_display::GitStatusDisplay,
     matched_string::MatchedString,
     screen::Screen,
+    cell_size::*,
 };
 use {
     crate::{
@@ -74,13 +76,13 @@ lazy_static! {
 pub const WIDE_STATUS: bool = true;
 
 /// the type used by all GUI writing functions
+//pub type W = std::io::BufWriter<std::io::Stderr>;
 pub type W = std::io::BufWriter<std::io::Stderr>;
 
 /// return the writer used by the application
 pub fn writer() -> W {
     std::io::BufWriter::new(std::io::stderr())
 }
-
 
 pub fn fill_bg(
     w: &mut W,